08_UI_csv_upload/requirements.txt
streamlit
pandas
htbuilder
streamlit_keycloak
openpyxl
psycopg2-binary
streamlit-date-picker
plotly
